Extract two tests to separate files
